backend/requirements.txt
Django==4.1.6
django-cors-headers==3.13.0
django-model-utils==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7.1
djangorestframework-simplejwt==5.2.2
django-filter==22.1
django-environ==0.9.0
boto3==1.26.95
pillow==9.5.0